+++ src/str.c
@@ -372,9 +372,17 @@ str_prepend( str *s, const char *addstr )
 		for ( i=s->len+lenaddstr-1; i>=lenaddstr; i-- )
 			s->data[i] = s->data[i-lenaddstr];
 	}
-	strncpy( s->data, addstr, lenaddstr );
+	// Georgi: fix the warning about truncation in strncpy
+	//   I am not sure why not use strcpy here, given that 'addstr' is a proper string,
+	//                                                       (see strlen(addstr) above)
+	// copy lenaddstr + 1 bytes to let strncpy put the terminating 0.
+	strncpy( s->data, addstr, lenaddstr + 1 );
 	s->len += lenaddstr;
-	s->data[ s->len ] = '\0';
+	// s->data[ s->len ] = '\0';
+
+	// strncpy( s->data, addstr, lenaddstr );
+	// s->len += lenaddstr;
+	// s->data[ s->len ] = '\0';
 }
 
 static inline void
@@ -392,7 +400,12 @@ str_strcat_internal( str *s, const char *addstr, unsigned long n )
 {
 	return_if_memerr( s );
 	str_strcat_ensurespace( s, n );
-	strncat( &(s->data[s->len]), addstr, n );
+	// Georgi: tell the compiler that there is enough space to fix
+	//   warning: 'strncat' output truncated before terminating nul copying as many bytes
+	//   from a string as its length [-Wstringop-truncation]
+	//
+	// strncat( &(s->data[s->len]), addstr, n );
+	strncat( &(s->data[s->len]), addstr, n + 1 - strlen(&(s->data[s->len])) - 1 );
 	s->len += n;
 	s->data[s->len]='\0';
 }
@@ -501,8 +514,17 @@ str_strcpy_internal( str *s, const char *p, unsigned long n )
 	return_if_memerr( s );
 
 	str_strcpy_ensurespace( s, n );
-	strncpy( s->data, p, n );
-	s->data[n] = '\0';
+	// Georgi: this fixes the warning about truncation in strncpy
+	//   strcpy cannot be used here since at least one of the calls below
+	//   passes a non-NULL terminated 'p'
+	//
+	// copy n + 1 bytes to let strncpy put the terminating 0.
+	strncpy( s->data, p, n + 1);
+	//s->data[n] = '\0';
+
+	// strncpy( s->data, p, n );
+	// s->data[n] = '\0';
+
 	s->len = n;
 }